Remove WPF dependency on the C.Core.Resource assembly
