+++ Cyclops.Core.Resource/Cyclops.Core.Resource.csproj
@@ -2,7 +2,6 @@
     <PropertyGroup>
         <TargetFramework>net6.0-windows</TargetFramework>
         <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-        <UseWpf>true</UseWpf>
         <UseWindowsForms>true</UseWindowsForms>
     </PropertyGroup>
     <ItemGroup>
